bonus/requirements.txt
numpy
pandas
Pillow
opencv-python
pytesseract
pdf2image